# io,layout: switch Source from struct to interface