@inkeliz 2024-05-22
I'm kindly pinging @eliasnaur since that issue is a significant for gio-plugins and other libraries. Could we consider either applying //go:noinline or using interface to address it?