@eliasnaur 2024-05-25
> gioexplorer.OpenFileOp{Tag: p.tag, Mimetype: FileTypes}.Add(gtx.Ops) // add custom event to Ops

If I understand this you're encoding a custom op into the internal `op.Ops.Internal` structure? This is not guaranteed to be stable.

How about calling `gioexplorer.Open(w.Invalidate)` so that a redraw can be triggered and the caller can check for incoming `transfer.DataEvent`s?