@inkeliz 2024-05-25
> If I understand this you're encoding a custom op into the internal op.Ops.Internal structure? This is not guaranteed to be stable.

That was how it works, before Gio 0.6. I know that it wasn't stable, and Gio, in general, isn't stable too. However, the goal was to make it similar to other kinds of Ops (like `clipboard.ReadOp`). 

I want to replace it with `gtx.Execute()`: `gtx.Execute(gioexplorer.Open{...})`. However, since `Source` isn't interface, I can't intercept that.

> How about calling gioexplorer.Open(w.Invalidate) so that a redraw can be triggered and the caller can check for incoming transfer.DataEvents?

But how I can send `transfer.DataEvents` into the Gio Router? Also, some plugins might require custom Events/Filters. For instance, `webviewer` sends events about cookies, javascript executions (...). That requires a custom Filter. The `Source`can only get Events, but can't send custom ones and the current code seems to use `switch` for filtering, which makes impossible to handle custom events/filters.

If `Source` is an interface, then I can use some kind of "decorator"/"wrapper" around the current `Source`. Or, if it uses `//go:noinline` (but that is hacky).