@inkeliz 2024-05-30
Benchmarks using gio-example kitchen:

## Current Patch (New = This Patch):

```
goos: darwin
goarch: arm64
pkg: gioui.org/example/kitchen
             │     old      │                new                 │
             │    sec/op    │    sec/op     vs base              │
UI-12          426.1µ ± ∞ ¹   441.3µ ± ∞ ¹  +3.58% (p=0.032 n=5)
UI_Offset-12   412.5µ ± ∞ ¹   416.8µ ± ∞ ¹  +1.04% (p=0.008 n=5)
UI_Scale-12    560.1µ ± ∞ ¹   555.4µ ± ∞ ¹       ~ (p=0.222 n=5)
UI_Rotate-12   670.5µ ± ∞ ¹   666.5µ ± ∞ ¹       ~ (p=0.151 n=5)
UI_All-12      672.5µ ± ∞ ¹   669.9µ ± ∞ ¹       ~ (p=0.690 n=5)
geomean        536.3µ         539.3µ        +0.55%
¹ need >= 6 samples for confidence interval at level 0.95

             │     old      │                 new                 │
             │    gc/op     │    gc/op      vs base               │
UI-12          13.80m ± ∞ ¹   15.50m ± ∞ ¹  +12.32% (p=0.008 n=5)
UI_Offset-12   13.11m ± ∞ ¹   14.65m ± ∞ ¹  +11.75% (p=0.008 n=5)
UI_Scale-12    18.90m ± ∞ ¹   20.70m ± ∞ ¹   +9.52% (p=0.008 n=5)
UI_Rotate-12   18.23m ± ∞ ¹   19.82m ± ∞ ¹   +8.72% (p=0.008 n=5)
UI_All-12      19.27m ± ∞ ¹   20.80m ± ∞ ¹   +7.94% (p=0.008 n=5)
geomean        16.44m         18.09m        +10.04%
¹ need >= 6 samples for confidence interval at level 0.95

             │     old      │                new                 │
             │  sec/frame   │  sec/frame    vs base              │
UI-12          391.3µ ± ∞ ¹   398.7µ ± ∞ ¹       ~ (p=0.056 n=5)
UI_Offset-12   377.3µ ± ∞ ¹   375.3µ ± ∞ ¹       ~ (p=0.841 n=5)
UI_Scale-12    525.5µ ± ∞ ¹   513.6µ ± ∞ ¹       ~ (p=0.151 n=5)
UI_Rotate-12   640.4µ ± ∞ ¹   627.7µ ± ∞ ¹  -1.98% (p=0.008 n=5)
UI_All-12      642.1µ ± ∞ ¹   630.8µ ± ∞ ¹  -1.76% (p=0.008 n=5)
geomean        502.1µ         497.3µ        -0.94%
¹ need >= 6 samples for confidence interval at level 0.95

             │     old      │                 new                 │
             │    sec/gc    │    sec/gc     vs base               │
UI-12          501.9n ± ∞ ¹   566.9n ± ∞ ¹  +12.95% (p=0.008 n=5)
UI_Offset-12   487.1n ± ∞ ¹   538.2n ± ∞ ¹  +10.49% (p=0.008 n=5)
UI_Scale-12    658.3n ± ∞ ¹   739.4n ± ∞ ¹  +12.32% (p=0.008 n=5)
UI_Rotate-12   676.1n ± ∞ ¹   744.5n ± ∞ ¹  +10.12% (p=0.008 n=5)
UI_All-12      748.8n ± ∞ ¹   764.7n ± ∞ ¹        ~ (p=0.095 n=5)
geomean        605.6n         663.3n         +9.53%
¹ need >= 6 samples for confidence interval at level 0.95

             │     old      │                 new                 │
             │  sec/layout  │  sec/layout   vs base               │
UI-12          34.49µ ± ∞ ¹   42.32µ ± ∞ ¹  +22.71% (p=0.008 n=5)
UI_Offset-12   35.01µ ± ∞ ¹   42.15µ ± ∞ ¹  +20.41% (p=0.008 n=5)
UI_Scale-12    34.27µ ± ∞ ¹   41.05µ ± ∞ ¹  +19.81% (p=0.008 n=5)
UI_Rotate-12   29.83µ ± ∞ ¹   38.36µ ± ∞ ¹  +28.59% (p=0.008 n=5)
UI_All-12      29.93µ ± ∞ ¹   38.51µ ± ∞ ¹  +28.66% (p=0.008 n=5)
geomean        32.62µ         40.44µ        +23.97%
¹ need >= 6 samples for confidence interval at level 0.95
```


-------------------

## No-Inline (New = //go:noinline version):

```
goos: darwin
goarch: arm64
pkg: gioui.org/example/kitchen
             │     old      │                new                 │
             │    sec/op    │    sec/op     vs base              │
UI-12          426.1µ ± ∞ ¹   427.9µ ± ∞ ¹       ~ (p=0.841 n=5)
UI_Offset-12   412.5µ ± ∞ ¹   402.1µ ± ∞ ¹  -2.51% (p=0.008 n=5)
UI_Scale-12    560.1µ ± ∞ ¹   529.0µ ± ∞ ¹  -5.56% (p=0.008 n=5)
UI_Rotate-12   670.5µ ± ∞ ¹   655.9µ ± ∞ ¹  -2.18% (p=0.008 n=5)
UI_All-12      672.5µ ± ∞ ¹   681.1µ ± ∞ ¹       ~ (p=0.095 n=5)
geomean        536.3µ         527.0µ        -1.74%
¹ need >= 6 samples for confidence interval at level 0.95

             │     old      │                new                 │
             │    gc/op     │    gc/op      vs base              │
UI-12          13.80m ± ∞ ¹   13.88m ± ∞ ¹       ~ (p=1.000 n=5)
UI_Offset-12   13.11m ± ∞ ¹   13.17m ± ∞ ¹       ~ (p=0.548 n=5)
UI_Scale-12    18.90m ± ∞ ¹   19.00m ± ∞ ¹       ~ (p=0.730 n=5)
UI_Rotate-12   18.23m ± ∞ ¹   18.11m ± ∞ ¹       ~ (p=0.643 n=5)
UI_All-12      19.27m ± ∞ ¹   19.15m ± ∞ ¹       ~ (p=0.127 n=5)
geomean        16.44m         16.45m        +0.06%
¹ need >= 6 samples for confidence interval at level 0.95

             │     old      │                new                 │
             │  sec/frame   │  sec/frame    vs base              │
UI-12          391.3µ ± ∞ ¹   392.7µ ± ∞ ¹       ~ (p=0.841 n=5)
UI_Offset-12   377.3µ ± ∞ ¹   367.8µ ± ∞ ¹  -2.51% (p=0.008 n=5)
UI_Scale-12    525.5µ ± ∞ ¹   496.0µ ± ∞ ¹  -5.61% (p=0.008 n=5)
UI_Rotate-12   640.4µ ± ∞ ¹   624.7µ ± ∞ ¹  -2.45% (p=0.008 n=5)
UI_All-12      642.1µ ± ∞ ¹   648.4µ ± ∞ ¹       ~ (p=0.151 n=5)
geomean        502.1µ         492.7µ        -1.88%
¹ need >= 6 samples for confidence interval at level 0.95

             │     old      │                 new                 │
             │    sec/gc    │    sec/gc     vs base               │
UI-12          501.9n ± ∞ ¹   552.1n ± ∞ ¹  +10.00% (p=0.008 n=5)
UI_Offset-12   487.1n ± ∞ ¹   511.0n ± ∞ ¹        ~ (p=0.056 n=5)
UI_Scale-12    658.3n ± ∞ ¹   704.9n ± ∞ ¹        ~ (p=0.056 n=5)
UI_Rotate-12   676.1n ± ∞ ¹   807.8n ± ∞ ¹  +19.48% (p=0.008 n=5)
UI_All-12      748.8n ± ∞ ¹   798.7n ± ∞ ¹   +6.66% (p=0.008 n=5)
geomean        605.6n         663.2n         +9.51%
¹ need >= 6 samples for confidence interval at level 0.95

             │     old      │                new                 │
             │  sec/layout  │  sec/layout   vs base              │
UI-12          34.49µ ± ∞ ¹   34.89µ ± ∞ ¹       ~ (p=1.000 n=5)
UI_Offset-12   35.01µ ± ∞ ¹   33.99µ ± ∞ ¹  -2.90% (p=0.008 n=5)
UI_Scale-12    34.27µ ± ∞ ¹   32.65µ ± ∞ ¹  -4.71% (p=0.008 n=5)
UI_Rotate-12   29.83µ ± ∞ ¹   30.87µ ± ∞ ¹  +3.50% (p=0.008 n=5)
UI_All-12      29.93µ ± ∞ ¹   30.79µ ± ∞ ¹  +2.85% (p=0.008 n=5)
geomean        32.62µ         32.60µ        -0.07%
¹ need >= 6 samples for confidence interval at level 0.95
```


--------